requirements.txt
click==7.1.2
h11==0.9.0
httptools==0.1.1
python-engineio==3.13.0
python-socketio==4.6.0
six==1.15.0
uvicorn==0.11.5
uvloop==0.14.0
websockets==8.1
